+++ src/services/Sso.php
@@ -33,9 +33,9 @@ class Sso extends Component
     // =========================================================================
 
     /**
-     * @event SsoDataEvent The event that is triggered before the SSO data is
-     *        used You may set [[SsoDataEvent::isValid]] to `false` to prevent
-     *        SSO data from being used.
+     * @event SsoDataEvent The event that is triggered before the SSO data is used,
+     * you may modify the [[SsoDataEvent::data]] as you see fit. You may set
+     * [[SsoDataEvent::isValid]] to `false` to prevent SSO data from being used.
      *
      * ```php
      * use nystudio107\vanillaforums\services\Sso;
@@ -75,14 +75,13 @@ class Sso extends Component
                 $request->get(),
                 $settings->vanillaForumsClientID,
                 $settings->vanillaForumsSecret,
-                true
+                $settings->hashAlgorithm ?? 'md5'
             );
             //$result = ob_get_contents();
             //ob_end_clean(); // Store buffer in variable
         }
 
         Craft::$app->end();
-
         //return $result === false ? '' : $result;
     }