+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # Vanillaforums Changelog
 
+## 1.0.1 - 2019.10.17
+### Added
+* Added a **Hash Algorithm** dropdown in Settings to allow you to choose the hash algorithm type
+* Added **Authentication URL** to the Settings page, which you can copy & paste into your Vanilla Forums settings
+
+### Changed
+* `vanillaForumsClientID` & `vanillaForumsSecret` are now validated via `EnvAttributeParserBehavior`
+* Updated the documentation to include information on the `SsoDataEvent` that Vanilla Forums throws
+ 
 ## 1.0.0 - 2019.10.16
 ### Added
 - Initial release